ptr size align

+++ my_malloc.c
@@ -1,4 +1,5 @@
 #include<stdlib.h>
+#include<inttypes.h>
 #include "my_malloc.h"
 
 int total_bytes_allocated = 0;
@@ -7,9 +8,9 @@ void *my_malloc(int size) {
     void *p;
 
     total_bytes_allocated += size;
-    p = malloc(size + sizeof(int));  // assuming that malloc returns word aligned pointers
-    ((int *) p)[0] = size;
-    p += sizeof(int);
+    p = malloc(size + sizeof(uint64_t));  // assuming that malloc returns word aligned pointers
+    ((uint64_t *) p)[0] = size;
+    p += sizeof(uint64_t);
     return (p);
 }
 
@@ -19,14 +20,14 @@ void *my_calloc(int num_elements, int element_size) {
     total_bytes_allocated += (num_elements * element_size);
 
     p = calloc(num_elements + 1, element_size);
-    ((int *) p)[0] = num_elements * element_size;
-    p += sizeof(int);
+    ((uint64_t *) p)[0] = num_elements * element_size;
+    p += sizeof(uint64_t);
     return (p);
 }
 
 void my_free(void *p) {
-    p -= sizeof(int);
-    int size = ((int *) p)[0];
+    p -= sizeof(uint64_t);
+    int size = ((uint64_t *) p)[0];
     total_bytes_allocated -= size;
     free(p);
 }
